Admin: Filter users by application signin permission
The signin permission was a fake boolean field which gets removed
and injected into the permissions array, having the value just in
the array makes it hard to filter by, so by duplicating the data
in a boolean we can filter the users who have signin permission.

Uses ransack to handle searching, and updated the exisiting search.

Does not delete the signin_permission param after submitting the user
permssion form, so it set as a boolean.

The motivation for this is to allow admins to filter user by
application to see who has access to what.

Adds a rake task to clean up the data so we can query the signon
property

    bundle exec rake users:fix_signin